fix(validator): compatible with py3
